pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "redis>=5.0.1",
    "python-json-logger>=2.0.7",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
]
